+++ Project/helpers/utils.ts
@@ -8,16 +8,35 @@ const EXPOSED_KEY =
   "8da4ef21b864d2cc526dbdb2a120bd2874c36c9d0a1fb7f8c63d7f7a8b41de8f";
 
 function getSigner(providerName: string = "goerli"): ethers.Wallet {
-    const wallet =
-      process.env.MNEMONIC && process.env.MNEMONIC.length > 0
-        ? ethers.Wallet.fromMnemonic(process.env.MNEMONIC)
-        : new ethers.Wallet(process.env.PRIVATE_KEY ?? EXPOSED_KEY);
-    
-        console.log(`Using address ${wallet.address}`);
-  
-    const provider = ethers.providers.getDefaultProvider(providerName);
-    
-    return wallet.connect(provider);
+  const wallet =
+    process.env.MNEMONIC && process.env.MNEMONIC.length > 0
+      ? ethers.Wallet.fromMnemonic(process.env.MNEMONIC)
+      : new ethers.Wallet(process.env.PRIVATE_KEY ?? EXPOSED_KEY);
+
+  console.log(`Using address ${wallet.address}`);
+
+  const provider = ethers.providers.getDefaultProvider(providerName);
+
+  return wallet.connect(provider);
 }
 
-  export { EXPOSED_KEY, getSigner };
+async function checkBalance(signer: ethers.Wallet): Promise<boolean> {
+  const balanceBN = await signer.getBalance();
+  const balance = Number(ethers.utils.formatEther(balanceBN));
+  console.log(`Wallet balance ${balance}`);
+  if (balance < 0.01) {
+    console.log("INSUFFICIENT ETH BALANCE");
+    return false;
+  }
+  return true;
+}
+
+function convertStringArrayToBytes32(array: string[]) {
+  const bytes32Array = [];
+  for (let index = 0; index < array.length; index++) {
+    bytes32Array.push(ethers.utils.formatBytes32String(array[index]));
+  }
+  return bytes32Array;
+}
+
+export { convertStringArrayToBytes32, getSigner, checkBalance };
