+++ Project/scripts/Ballot/queryProposals.ts
@@ -3,33 +3,15 @@ import "dotenv/config";
 import * as ballotJson from "../../artifacts/contracts/Ballot.sol/Ballot.json";
 // eslint-disable-next-line node/no-missing-import
 import { Ballot } from "../../typechain";
-// This key is already public on Herong's Tutorial Examples - v1.03, by Dr. Herong Yang
-// Do never expose your keys like this
-
-// DEPLOYS TO GOERLI, NOT HARDHAT SIM!!
-
-const EXPOSED_KEY =
-  "8da4ef21b864d2cc526dbdb2a120bd2874c36c9d0a1fb7f8c63d7f7a8b41de8f";
-
-function convertStringArrayToBytes32(array: string[]) {
-  const bytes32Array = [];
-  for (let index = 0; index < array.length; index++) {
-    bytes32Array.push(ethers.utils.formatBytes32String(array[index]));
-  }
-  return bytes32Array;
-}
+import { checkBalance, getSigner } from "../../helpers/utils";
 
 async function main() {
-  const wallet =
-    process.env.MNEMONIC && process.env.MNEMONIC.length > 0
-      ? ethers.Wallet.fromMnemonic(process.env.MNEMONIC)
-      : new ethers.Wallet(process.env.PRIVATE_KEY ?? EXPOSED_KEY);
-  console.log(`Using address ${wallet.address}`);
-  const provider = ethers.providers.getDefaultProvider("goerli");
-  const signer = wallet.connect(provider);
-  const balanceBN = await signer.getBalance();
-  const balance = Number(ethers.utils.formatEther(balanceBN));
-  console.log(`Wallet balance ${balance}`);
+  const signer = getSigner();
+
+  const hasSufficientBalance = await checkBalance(signer);
+  if (!hasSufficientBalance) {
+    return;
+  }
   if (process.argv.length < 3) throw new Error("Ballot address missing");
   const ballotAddress = process.argv[2];
 
@@ -42,7 +24,7 @@ async function main() {
   console.log("Proposals: ");
   const numProposals = 5;
   for (let i = 0; i < numProposals; i++) {
-    var proposal = ethers.utils.parseBytes32String(
+    let proposal = ethers.utils.parseBytes32String(
       (await ballotContract.proposals(i)).name
     );
     console.log(`Proposal N. ${i + 1}: ${proposal}`);
