+++ Project/tsconfig.json
@@ -8,6 +8,6 @@
     "declaration": true,
     "resolveJsonModule": true,
   },
-  "include": ["./scripts", "./tests", "./typechain, ./helpers"],
+  "include": ["./scripts", "./tests", "./typechain", "./helpers"],
   "files": ["./hardhat.config.ts"]
 }
